requirements.txt
selenium
pandas
python-dotenv
bs4
requests